Revert "change include directives to use double quotes"
This does not seem to be a complete and proper fix for the conflict
between uclibc++ new/delete and the ones defined by the AVR core, since
it only fixes conflicts in the header files, but source files might
still produce linker errors.

So better revert it and stay closer to upstream uclibc++ sources.

This reverts commit 84b4a911b8228bf3dfb4133f3992ddad36366a3a.